fix: add authorization check to delete messages endpoint

+++ app/routes/api.delete-messages.tsx
@@ -1,12 +1,20 @@
 import { deleteExpiredOrOneTimeMessages } from "@/.server/message";
-import { json } from "@vercel/remix";
+import { json, LoaderFunctionArgs } from "@vercel/remix";
 
-export async function loader() {
+export async function loader({ request }: LoaderFunctionArgs) {
+  const authHeader = request.headers.get("authorization");
+  if (
+    process.env.NODE_ENV !== "development" &&
+    authHeader !== `Bearer ${process.env.CRON_SECRET}`
+  ) {
+    return new Response("Unauthorized", {
+      status: 401,
+    });
+  }
   try {
     await deleteExpiredOrOneTimeMessages();
     return json("Messages deleted successfully.");
   } catch (error) {
-    console.log(error);
     return json("An error occurred while attempting to delete messages.", {
       status: 500,
     });
